requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
python-dotenv
joblib
email-validator